pyproject.toml
[project]
name = "rpaframework-workitems"
version = "29.0.0"
description = ""
readme = "README.md"
authors = [
    {name = "Asko Soukka", email = "asko.soukka@iki.fi"},
]
license = {text = "Apache Software License" }
keywords = ["robotframework", "rpa", "automation"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Framework :: Robot Framework :: Library",
    "Framework :: Robot Framework",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">= 3.9"
dependencies = [
    "robotframework>=7.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/RPA"]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
]

[tool.isort]
force_alphabetical_sort_within_sections = true
force_single_line = true
from_first = true
line_length = 200
lines_after_imports = 2
lines_between_types = 0
no_sections = true
